diary/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pydantic
passlib[bcrypt]
python-jose[cryptography]
fastapi-mail
authlib
requests
